+++ CHANGELOG.md
@@ -168,3 +168,4 @@
 - Improvements to `ObjectTree`
   - Fix an inefficiency in sequence traversal
   - Fix a bug when `skipNones` was `false`
+  - Add a maximum recursion depth