+++ CHANGELOG.md
@@ -170,3 +170,7 @@
   - Fix a bug when `skipNones` was `false`
   - Add a maximum recursion depth
   - Add a length limit for single-line 1-products
+
+#### 4.0.2
+
+- Update SBT to 1.1.4